+++ src/components/atoms/RoundedButton/RoundedButton.tsx
@@ -2,14 +2,14 @@ import Community from './Community';
 import Home from './Home';
 import Analysis from './Analysis';
 import Play from './Play';
-import RoutineUpdate from './RoutineUpdate';
+import Edit from './Edit';
 
 const RoundedButton = {
   Community,
   Home,
   Analysis,
   Play,
-  RoutineUpdate,
+  Edit,
 };
 
 export default RoundedButton;
